Switched readInt( to readUnsignednt( due to player addresses being larger than a normal signed int

+++ src/main/java/org/xena/cs/Addressed.java
@@ -2,8 +2,8 @@ package org.xena.cs;
 
 public interface Addressed {
 
-    int address();
+    long address();
 
-    void setAddress(int address);
+    void setAddress(long address);
 
 }
